+++ src/vscode/filesystemProvider/fabric/FabricFSWorkspace.ts
@@ -1,14 +1,10 @@
 import * as vscode from 'vscode';
 
-import { ThisExtension } from '../../../ThisExtension';
-import { Helper } from '../../../helpers/Helper';
-import { FabricApiItemType, FabricApiWorkspaceType, iFabricApiWorkspace } from '../../../fabric/_types';
-import { FABRIC_FS_ITEM_TYPES } from './_types';
+import { FabricApiWorkspaceType, iFabricApiWorkspace } from '../../../fabric/_types';
 import { FabricFSCacheItem } from './FabricFSCacheItem';
 import { FabricFSUri } from './FabricFSUri';
 import { FabricApiService } from '../../../fabric/FabricApiService';
-
-
+import { PowerBIConfiguration } from '../../configuration/PowerBIConfiguration';
 
 export class FabricFSWorkspace extends FabricFSCacheItem implements iFabricApiWorkspace {
 	id: string;
@@ -48,9 +44,9 @@ export class FabricFSWorkspace extends FabricFSCacheItem implements iFabricApiWo
 	public async loadChildrenFromApi<T>(): Promise<void> {
 		if (!this._children) {
 			this._children = [];
-			this._apiResponse = FABRIC_FS_ITEM_TYPES;
-			for (let itemType of FABRIC_FS_ITEM_TYPES) {
-				this._children.push([FabricApiItemType[itemType], vscode.FileType.Directory]);
+			this._apiResponse = PowerBIConfiguration.fabricItemTypeNames;
+			for (let itemType of this._apiResponse) {
+				this._children.push([itemType, vscode.FileType.Directory]);
 			}
 		}
 	}